:bug: update launch script

+++ main.sh
@@ -1,26 +1,48 @@
 #!/usr/bin/env bash
+#SBATCH -N 2
+#SBATCH -n 32
+#SBATCH --mem-per-cpu=1024
+# source scl_source enable gcc-toolset-11
+# module load hpcx-2.7.0/hpcx-ompi
+# module load openmpi/4.1.5
 src="graph-openmp"
 out="$HOME/Logs/$src.log"
 ulimit -s unlimited
 printf "" > "$out"
 
 # Download program
-rm -rf $src
-git clone https://github.com/ionicf/$src
+if [[ "$DOWNLOAD" != "0" ]]; then
+  rm -rf $src
+  git clone https://github.com/puzzlef/$src
+fi
 cd $src
 
+# Fixed config
+: "${TYPE:=double}"
+: "${MAX_THREADS:=32}"
+: "${REPEAT_BATCH:=5}"
+: "${REPEAT_METHOD:=1}"
+# Define macros (dont forget to add here)
+DEFINES=(""
+"-DTYPE=$TYPE"
+"-DMAX_THREADS=$MAX_THREADS"
+"-DREPEAT_BATCH=$REPEAT_BATCH"
+"-DREPEAT_METHOD=$REPEAT_METHOD"
+)
+
 # Run
-g++ -std=c++17 -O3 -fopenmp main.cxx
-stdbuf --output=L ./a.out ~/Data/GAP-road.mtx       2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/GAP-twitter.mtx    2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/GAP-web.mtx        2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/GAP-kron.mtx       2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/GAP-urand.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/com-Orkut.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/com-Friendster.mtx 2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/arabic-2005.mtx    2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/uk-2005.mtx        2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/webbase-2001.mtx   2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/it-2004.mtx        2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/sk-2005.mtx        2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/Data/kmer_V1r.mtx       2>&1 | tee -a "$out"
+g++ ${DEFINES[*]} -std=c++17 -O3 -fopenmp main.cxx
+stdbuf --output=L ./a.out ~/Data/soc-Epinions1.mtx   2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/indochina-2004.mtx  2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/uk-2002.mtx         2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/arabic-2005.mtx     2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/uk-2005.mtx         2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/webbase-2001.mtx    2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/it-2004.mtx         2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/sk-2005.mtx         2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/com-LiveJournal.mtx 2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/com-Orkut.mtx       2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/asia_osm.mtx        2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/europe_osm.mtx      2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/kmer_A2a.mtx        2>&1 | tee -a "$out"
+# stdbuf --output=L ./a.out ~/Data/kmer_V1r.mtx        2>&1 | tee -a "$out"
